Move tests of css linter to 'analyzer.test.ts'
